requirements.txt
Flask>=3.0
SQLAlchemy>=2.0
APScheduler>=3.10
requests
python-dotenv
pyttsx3           # TTS
vosk              # STT（モデルは手動DL）
SpeechRecognition  # STT (マイク入力用)
sounddevice      # Audio I/O for Vosk
msal              # Microsoft Graph 認証
pandas            # Excel 操作補助
openpyxl          # ↑のドライバ
pytest            # テスト用
playsound==1.2.2
